Adds pam_environment for gpg_agent activation

+++ bin/install.sh
@@ -3,6 +3,7 @@
 set -euo pipefail
 
 cd dotfiles/home
-ln --backup -s $(pwd)/bashrc ~/.bashrc
-ln --backup -s $(pwd)/emacs ~/.emacs
+[[ -L ~/.bashrc ]] || ln --backup -s $(pwd)/bashrc ~/.bashrc
+[[ -L ~/.emacs ]] || ln --backup -s $(pwd)/emacs ~/.emacs
+[[ -L ~/.pam_environment ]] || ln --backup -s $(pwd)/pam_environment ~/.pam_environment
 cd -
